fix(comms/dht)!: limit number of peer claims and addresses for all sources (#5702)
Description
---
Implement validations and bans for peers that violate peer identity
exchange protocol rules
Consolidate peer claim and identity validations
Removed a number of dead functions and error variants
Reduced memory footprint of peer connection handle
Removed unverified claim data, this data is not unverified and is only
ever added for authenticated direct connections
Check that discovery response comes from the peer that directly
connected to us
Reduce peer flags size from u64 to u32 (smaller would require many more
changes)
Reduce memory footprint in dialer

Motivation and Context
---
TARI-021

All peer source validations use the same code. We previously allowed a
peer to send any number of identity claims when doing discovery.
Discovery RPC protocol now allows a limited number of claims to be sent,
ordered by local quality score.

How Has This Been Tested?
---
Existing test, localnet network (2 bn 2 wallets)

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [ ] None
- [x] Requires _peer_ data directory on base node to be deleted
- [ ] Requires hard fork
- [x] Other - some network messages relating to peer sync are not
compatible
<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

+++ base_layer/wallet/src/storage/sqlite_db/wallet.rs
@@ -303,7 +303,7 @@ impl WalletSqliteDatabase {
 
     fn get_comms_features(&self, conn: &mut SqliteConnection) -> Result<Option<PeerFeatures>, WalletStorageError> {
         if let Some(key_str) = WalletSettingSql::get(&DbKey::CommsFeatures, conn)? {
-            let features = u64::from_str(&key_str).map_err(|e| WalletStorageError::ConversionError(e.to_string()))?;
+            let features = u32::from_str(&key_str).map_err(|e| WalletStorageError::ConversionError(e.to_string()))?;
             let peer_features = PeerFeatures::from_bits(features);
             Ok(peer_features)
         } else {
